* Removed some unused variables. * Fixed some casting warnings.
git-svn-id: http://winbolo.googlecode.com/svn/trunk@174 f5f8f8c0-f940-0410-bb6b-693536db5c03

+++ winbolo/src/bolo/gametype.c
@@ -103,7 +103,7 @@ void gameTypeGetItems(gameType *gmeType, BYTE *shellsAmount, BYTE *mines, BYTE *
     if (numBases == 0) {
       numBases = 1;
     }
-    percent = screenGetNumNeutralBases();
+    percent = (double) screenGetNumNeutralBases();
     percent /= numBases;
     percent *= BASE_MAX_DOUBLE;
     *shellsAmount = (BYTE) (TWO * percent);
